[FE] REFACTOR: 폴더 구조 조정 #123
