+++ src/frontend/src/components/Sidebar/Chating/ChatInput/index.css.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const InputContainer = styled.div`
+  display: flex;
+  align-items: center;
+  padding: 8px;
+  border-top: 1px solid #d4d4d4;
+  background: white;
+`;
